@@ -37,13 +37,16 @@
     "jscs": "^3.0.7",
     "less": "^3.7.0",
     "less-loader": "^2.0.0",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2",
     "style-loader": "^0.13.1",
     "superagent": "^3.8.3",
     "url-loader": "^1.0.1",
     "webpack": "^1.12.14",
     "webpack-dev-middleware": "^1.6.0",
     "webpack-hot-middleware": "^2.10.0"
+  },
+  "devDependencies": {
+    "react-bootstrap": "^0.32.1"
   }
 }
