@@ -20,6 +20,7 @@ class HomepagePresenter extends Nette\Application\UI\Presenter
 				'brand' => $row->vendor,
 				'model' => $row->model,
 				'year' => $row->year,
+				'km' => $row->km,
 				'details' => FALSE,
 			];
 		}
@@ -28,11 +29,12 @@ class HomepagePresenter extends Nette\Application\UI\Presenter
 		$this->presenter->sendPayload();
 	}
 
-	public function actionAddCar($vendor, $model, $year) {
+	public function actionAddCar($vendor, $model, $year, $km) {
 		$row = $this->context->table('car')->insert([
 			'vendor' => $vendor,
 			'model' => $model,
 			'year' => $year,
+			'km' => $km,
 		]);
 		$this->presenter->payload->id = $row->id;
 		$this->presenter->sendPayload();
@@ -41,6 +43,19 @@ class HomepagePresenter extends Nette\Application\UI\Presenter
 
 	public function actionRemoveCar($id) {
 		$this->context->table('car')->where('id', $_GET['id'])->delete();
+		$this->presenter->payload->result = TRUE;
+		$this->presenter->sendPayload();
+	}
+
+	public function actionEditCar($id, $vendor, $model, $year, $km) {
+		$this->context->table('car')
+			->where('id', $_GET['id'])
+			->update([
+				'vendor' => $vendor,
+				'model' => $model,
+				'year' => $year,
+				'km' => $km,
+		]);
 		$this->presenter->payload->result = TRUE;
 		$this->presenter->sendPayload();
 	}
