@@ -1,8 +1,9 @@
 import React from 'react';
 import Car from './car';
 import AddCar from './addCar';
-import request from "superagent";
+import EditCar from './addCar';
 import ApiClient from './api/ApiClient';
+import BaseModal from './baseModal';
 
 export default class Cars extends React.Component {
 
@@ -14,28 +15,61 @@ export default class Cars extends React.Component {
         this.addCar = this.addCar.bind(this);
         this.removeCar = this.removeCar.bind(this);
         this.toggleCar = this.toggleCar.bind(this);
+        this.editCar = this.editCar.bind(this);
+        this.saveModalState = this.saveModalState.bind(this);
+        this.closeModal = this.closeModal.bind(this);
         let self = this;
         this.apiClient.findCars(function(cars) {
             self.setState(cars);
             self.render();
-
         });
-        this.state = {cars: []};
+        this.state = {
+            cars: [],
+            modal: {
+                title: 'vychozi',
+                show: false
+            },
+        };
     }
 
     render() {
         return (
             <div>
+                <BaseModal
+                    title={this.state.modal.title}
+                    show={this.state.modal.show}
+                    onClose={this.closeModal}
+                >
+                    <EditCar add={this.addCar} car={this.state.modal.car}/>
+                </BaseModal>
                 <AddCar add={this.addCar} />
                 <h3>Auta ({this.state.cars.length})</h3>
                 {this.state.cars.map(car => {
                     return <Car car={car}
                                 key={car.id}
                                 remove={this.removeCar}
-                                toggle={this.toggleCar} />;
+                                toggle={this.toggleCar}
+                                edit={this.editCar}
+                    />;
                 })}
             </div>
         );
+    }
+
+    closeModal() {
+        this.setState({
+            modal: {
+                show: false
+            }
+        });
+    }
+
+    saveModalState() {
+        console.log('save modal state');
+    }
+
+    editCar(carId, brand, model, year, km) {
+
     }
 
     addCar(car) {
@@ -45,7 +79,6 @@ export default class Cars extends React.Component {
         this.setState({
             cars: this.state.cars.concat([car])
         });
-        this.apiClient.addCar(car.brand, car.model, car.year);
     }
 
     removeCar(id) {
@@ -65,6 +98,18 @@ export default class Cars extends React.Component {
                 return car;
             })
         });
+    }
+
+    editCar(car) {
+        // @todo udelat to stejne jako je udelane toggle
+        this.setState({
+            modal: {
+                title: 'Editace auta ' + car.brand,
+                show: true,
+                car: car
+            }
+        });
+        console.log('state po edit car', this.state);
     }
 
 }
